Add rustls feature flag
Adds a non-default feature flag "rustls" to build against rustls instead
of native-tls.

Native-tls wraps the OpenSSL crate which is at times incompatible with
the current version of LibreSSL, e.g., on OpenBSD -current.

Rustls is a pure Rust TLS library that does not depend on OpenSSL.

Tested on OpenBSD -current with LibreSSL 3.5.2 and xterm.