ci: switch from actions-rs to dtolnay for clippy/rustfmt
`actions-rs` is largely unmaintained for a few years. In contrast,
dtolnay's actions are maintained and have largely taken the place of
actions-rs for many Rust workflows.

This patch switches to dtolnay's versions for Rustfmt and Clippy, but
does not touch the other jobs since they are more complex with matrix
testing. Note that this also changes from running nightly clippy to
stable clippy, which should pass, unlike what we see right now [[1]].

[1]: https://github.com/hatoo/oha/actions/runs/6281404969/job/17059738170