Add support for fractional QPS values
This adds support for fractional QPS values (QPS is no longer required to be an integer).
There is a new test using a fractional 1.75 QPS value, the test would fail if the QPS value
would be treated as integer.

There's also a small refactoring in the tests to create a new `test_request_count` function
to enable using the same function in all of the query limit tests.